refactor: project store tests
